split the logic of moving, victory checking as suggested by realiarthur
